+++ Vault-Coder/Changes.md
@@ -15,6 +15,7 @@
 ### Bugfixes
 - Optional references didn't work previously - now fixed
 - Fixed a bug where SQL indices where written incorrectly
+- Fixed a bug in DbModel / Factory writing where wrong property names were used
 - Fixed a bug in some factory implementations where read values were not unwrapped correctly
   - Please note that as a side effect, **ValueUnwraps** are no longer used when parsing instance data from models
 ### Other Changes